+++ src/components/_private/Flyout/FlyoutControlled.tsx
@@ -127,6 +127,7 @@ const FlyoutRoot = (props: T.ControlledProps & T.DefaultProps) => {
 	}, [clearTimer, timerRef, handleClose]);
 
 	const handleTriggerClick = React.useCallback(() => {
+		console.log("HEHEHEHH", status);
 		if (status === "idle") {
 			handleOpen();
 		} else {