+++ src/components/_private/Portal/Portal.types.ts
@@ -2,7 +2,13 @@ import React from "react";
 
 export type Props = {
 	children?: React.ReactNode;
-	scopeRef?: React.RefObject<HTMLElement | null>;
+	targetRef?: React.RefObject<HTMLElement | null>;
 };
 
-export type Context = Pick<Props, "scopeRef">;
+export type ScopeProps<T extends HTMLElement> = {
+	children: (ref: React.RefObject<T>) => React.ReactNode;
+};
+
+export type Context = {
+	scopeRef: React.RefObject<HTMLElement | null>;
+};