+++ src/components/_private/Portal/index.ts
@@ -1,2 +1,2 @@
-export { default, usePortal } from "./Portal";
+export { default } from "./Portal";
 export type { Props as PortalProps } from "./Portal.types";