+++ src/components/_private/Portal/tests/Portal.stories.tsx
@@ -4,20 +4,20 @@ import Portal from "components/_private/Portal";
 export default { title: "Utilities/Internal/Portal" };
 
 export const base = () => {
-	const ref = React.useRef<HTMLDivElement>(null);
+	const ref = React.useRef<HTMLDivElement | null>(null);
+	const [mounted, setMounted] = React.useState(false);
+
+	React.useEffect(() => {
+		setMounted(true);
+	}, []);
 
 	return (
-		<div style={{ border: "2px solid #333", padding: 16 }}>
-			App
-			<Portal scopeRef={ref}>
-				<div ref={ref} style={{ border: "2px solid green", padding: 16 }}>
-					Portal scope
-					<div style={{ border: "2px solid red", padding: 16 }}>
-						Stays in red
-						<Portal>Ported to green</Portal>
-					</div>
-				</div>
-			</Portal>
-		</div>
+		<>
+			<div style={{ border: "2px solid #333", padding: 16 }}>
+				App
+				{mounted && <Portal targetRef={ref}>Ported to green</Portal>}
+			</div>
+			<div ref={ref} style={{ border: "2px solid green", padding: 16 }} />
+		</>
 	);
 };