+++ CHANGELOG.md
@@ -4,3 +4,6 @@
 
 - Remove tap-highlight for checkbox, radio and switch on mobile devices
 - Select: Fixed disabled icon color
+
+- Actionable: Fixed repeating onClicks for Enter
+- Popover, DropdownMenu: Added rendering support inside scrollable areas